There is now a fully functional hotbar
The player can now pick up items and place them in the hotbar. The hot
bar is scrolled through with the scroll wheele. The player can also
place different blocks based on what is selected in the hotbar. There is
also a sound when the player places a block now.